#723 Testing the search

+++ ontrack-model/src/main/java/net/nemerosa/ontrack/model/structure/SearchResult.java
@@ -20,6 +20,6 @@ public class SearchResult {
      */
     private final URI page;
 
-    private final int accuracy;
+    private final double accuracy;
 
 }
